requirements.txt
Flask==2.3.3
boto3==1.28.85
python-dotenv==1.0.0
botocore==1.31.85
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0